+++ src/services.py
@@ -17,6 +17,7 @@ transactions = pd.read_excel(file_path)
 
 
 def safe_lower(value):
+    """Безопасное приведение строки к нижнему регистру."""
     if pd.isna(value):
         return ''
     return value.lower()
@@ -26,21 +27,23 @@ def search_transactions(query: str) -> str:
     """Поиск транзакций по строке."""
     logger.info(f"Поиск транзакций по запросу: {query}")
     result = list(
-        filter(lambda t: query.lower() in safe_lower(t['Описание']) or query.lower() in safe_lower(t['Категория']),
-               transactions.to_dict(orient='records')))
+    filter(lambda t: (query.lower() in safe_lower(t['Описание']) if t['Описание'] else False) or
+                  (query.lower() in safe_lower(t['Категория']) if t['Категория'] else False), transactions)
+)
     return json.dumps(result, ensure_ascii=False)
 
 
-def search_phone_numbers() -> str:
+def search_phone_numbers(transactions: pd.DataFrame) -> str:
     """Поиск транзакций по телефонным номерам."""
     logger.info("Поиск транзакций по телефонным номерам")
     phone_pattern = r'\+?\d[\d\s-]{7,}'
-    result = list(
-        filter(lambda t: re.search(phone_pattern, safe_lower(t['Описание'])), transactions.to_dict(orient='records')))
+    transactions_list = transactions.to_dict(orient='records')
+    result = list(filter(lambda t: t['Описание'] is not None and re.search(phone_pattern, t['Описание'].lower()),
+                         transactions_list))
     return json.dumps(result, ensure_ascii=False)
 
 
-def search_personal_transfers() -> str:
+def search_personal_transfers(transactions: pd.DataFrame) -> str:
     """Поиск переводов физическим лицам."""
     logger.info("Поиск переводов физическим лицам")
     personal_transfer_pattern = r'[А-Яа-я]\s+[А-Яа-я]\.'
